requirements.txt
selenium-wire==4.11.2
pandas==2.0.3
flask==2.0.1
requests==2.26.0
fake-useragent==0.1.11
pysocks==1.7.1
gunicorn==20.1.0
webdriver-manager==3.8.5
cloudscraper==1.2.71 